% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ie}
\alias{ie}
\title{Interval estimation data reported in Vuorre & Metcalfe (2017)}
\format{A data frame (\code{tibble::tibble}) with 3011 rows and 5 variables:
\describe{
  \item{id}{ID of study participant.}
  \item{condition}{Experimental condition.}
  \item{interval}{Inter-stimulus interval.}
  \item{estimate}{Numerical duration estimate (participant's response).}
  \item{cond}{Dummy coded condition, 0=involuntary, 1=voluntary.}
}}
\source{
Vuorre, M. & Metcalfe, J., (2017).
Voluntary action alters the perception of visual illusions.
Attention, Perception, & Psychophysics
\url{https://doi.org/10.3758/s13414-017-1321-x}
}
\usage{
data(ie)
}
\description{
Interval estimation data reported in Vuorre & Metcalfe (2017)
}
\keyword{datasets}
