% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illusion}
\alias{illusion}
\title{Visual illusion data reported in Vuorre & Metcalfe (2017)}
\format{A data frame (\code{tibble::tibble}) with 60,872 rows and 7 variables:
\describe{
  \item{exp}{Experiment number.}
  \item{id}{ID of study participant.}
  \item{exclude}{Was this data excluded from analysis.}
  \item{condition}{Experimental condition.}
  \item{cond}{Dummy coded condition. (0=control cond, 1=exp.)}
  \item{interval}{Inter-stimulus interval.}
  \item{response}{0 = no motion (exp 1) or group motion (exp 2-4),
  1 = motion (exp 1) or element motion (exp 2-4)}
}}
\source{
Vuorre, M. & Metcalfe, J., (2017).
Voluntary action alters the perception of visual illusions.
Attention, Perception, & Psychophysics
\url{https://doi.org/10.3758/s13414-017-1321-x}
}
\usage{
data(illusion)
}
\description{
Visual illusion data reported in Vuorre & Metcalfe (2017)
}
\keyword{datasets}
